# swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080/api",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/payment/refund/{orderId}" : {
      "post" : {
        "tags" : [ "支付管理" ],
        "summary" : "申请退款",
        "description" : "对指定订单申请退款",
        "operationId" : "refundPayment",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/payment/create" : {
      "post" : {
        "tags" : [ "支付管理" ],
        "summary" : "创建支付",
        "description" : "为订单创建支付记录",
        "operationId" : "createPayment",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PaymentDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultPaymentResultDTO"
                }
              }
            }
          }
        }
      }
    },
    "/payment/callback" : {
      "post" : {
        "tags" : [ "支付管理" ],
        "summary" : "支付回调",
        "description" : "接收第三方支付平台的回调通知",
        "operationId" : "processPaymentCallback",
        "parameters" : [ {
          "name" : "paymentNumber",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/order/pay/{orderId}" : {
      "post" : {
        "tags" : [ "订单管理" ],
        "summary" : "支付订单",
        "description" : "对指定订单进行支付",
        "operationId" : "payOrder",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/order/create" : {
      "post" : {
        "tags" : [ "订单管理" ],
        "summary" : "创建订单",
        "description" : "创建新的车票订单",
        "operationId" : "createOrder",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OrderCreateDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultOrderInfoDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/order/cancel/{orderId}" : {
      "post" : {
        "tags" : [ "订单管理" ],
        "summary" : "取消订单",
        "description" : "取消指定订单ID的订单",
        "operationId" : "cancelOrder",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/payment/refund/check/{orderId}" : {
      "get" : {
        "tags" : [ "支付管理" ],
        "summary" : "退款检查",
        "description" : "检查订单是否可退及退款金额",
        "operationId" : "refundCheck",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultRefundCheckResult"
                }
              }
            }
          }
        }
      }
    },
    "/payment/order/{orderId}" : {
      "get" : {
        "tags" : [ "支付管理" ],
        "summary" : "根据订单ID查询支付",
        "description" : "通过订单ID获取支付记录",
        "operationId" : "getPaymentByOrderId",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultPaymentResultDTO"
                }
              }
            }
          }
        }
      }
    },
    "/payment/number/{paymentNumber}" : {
      "get" : {
        "tags" : [ "支付管理" ],
        "summary" : "根据支付号查询",
        "description" : "通过支付号精确查询支付记录",
        "operationId" : "getPaymentByNumber",
        "parameters" : [ {
          "name" : "paymentNumber",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultPaymentResultDTO"
                }
              }
            }
          }
        }
      }
    },
    "/order/{orderId}" : {
      "get" : {
        "tags" : [ "订单管理" ],
        "summary" : "根据ID查询订单",
        "description" : "通过订单ID获取订单详细信息",
        "operationId" : "getOrderById",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultOrderInfoDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/order/status/{status}" : {
      "get" : {
        "tags" : [ "订单管理" ],
        "summary" : "根据状态查询订单",
        "description" : "根据订单状态（待支付、已取消等）筛选订单",
        "operationId" : "getOrdersByStatus",
        "parameters" : [ {
          "name" : "status",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultListOrderInfoDTO"
                }
              }
            }
          }
        }
      }
    },
    "/order/number/{orderNumber}" : {
      "get" : {
        "tags" : [ "订单管理" ],
        "summary" : "根据订单号查询",
        "description" : "通过订单号精确查询订单",
        "operationId" : "getOrderByNumber",
        "parameters" : [ {
          "name" : "orderNumber",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultOrderInfoDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ResultBoolean" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "PaymentDTO" : {
        "required" : [ "orderId", "paymentAmount", "paymentMethod" ],
        "type" : "object",
        "properties" : {
          "orderId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "paymentAmount" : {
            "type" : "number"
          },
          "paymentMethod" : {
            "type" : "string"
          },
          "paymentChannel" : {
            "type" : "string"
          },
          "remark" : {
            "type" : "string"
          }
        }
      },
      "PaymentResultDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "orderId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "orderNumber" : {
            "type" : "string"
          },
          "paymentNumber" : {
            "type" : "string"
          },
          "paymentAmount" : {
            "type" : "number"
          },
          "paymentAmountText" : {
            "type" : "string"
          },
          "orderAmount" : {
            "type" : "number"
          },
          "orderAmountText" : {
            "type" : "string"
          },
          "paymentMethod" : {
            "type" : "string"
          },
          "paymentMethodText" : {
            "type" : "string"
          },
          "paymentChannel" : {
            "type" : "string"
          },
          "paymentStatus" : {
            "type" : "integer",
            "format" : "int32"
          },
          "paymentStatusText" : {
            "type" : "string"
          },
          "paymentTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updateTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
      "ResultPaymentResultDTO" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PaymentResultDTO"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "OrderCreateDTO" : {
        "required" : [ "arrivalStationId", "departureDate", "departureStationId", "passengerIdCard", "passengerRealName", "seatType", "trainId" ],
        "type" : "object",
        "properties" : {
          "trainId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "departureStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "arrivalStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "departureDate" : {
            "type" : "string",
            "format" : "date"
          },
          "seatType" : {
            "type" : "string"
          },
          "passengerIdCard" : {
            "pattern" : "\\d{17}[\\d|x|X]|\\d{15}",
            "type" : "string"
          },
          "passengerRealName" : {
            "type" : "string"
          },
          "ticketCount" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "OrderInfoDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "orderNumber" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "username" : {
            "type" : "string"
          },
          "trainId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "trainNumber" : {
            "type" : "string"
          },
          "departureStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "departureStationName" : {
            "type" : "string"
          },
          "arrivalStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "arrivalStationName" : {
            "type" : "string"
          },
          "departureDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "departureTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "arrivalTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "seatType" : {
            "type" : "string"
          },
          "seatTypeName" : {
            "type" : "string"
          },
          "passengerIdCard" : {
            "type" : "string"
          },
          "passengerRealName" : {
            "type" : "string"
          },
          "ticketPrice" : {
            "type" : "number"
          },
          "ticketPriceText" : {
            "type" : "string"
          },
          "orderStatus" : {
            "type" : "integer",
            "format" : "int32"
          },
          "orderStatusText" : {
            "type" : "string"
          },
          "payStatus" : {
            "type" : "integer",
            "format" : "int32"
          },
          "payStatusText" : {
            "type" : "string"
          },
          "payTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expireTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updateTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expired" : {
            "type" : "boolean"
          },
          "remainingPayTime" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "ResultOrderInfoDTO" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/OrderInfoDTO"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "NonRefundableTicket" : {
        "type" : "object",
        "properties" : {
          "ticketNumber" : {
            "type" : "string"
          },
          "passengerName" : {
            "type" : "string"
          },
          "trainNumber" : {
            "type" : "string"
          },
          "departureTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "seatInfo" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "suggestion" : {
            "type" : "string"
          },
          "displayInfo" : {
            "type" : "string"
          }
        }
      },
      "RefundCheckResult" : {
        "type" : "object",
        "properties" : {
          "refundable" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "checkId" : {
            "type" : "string"
          },
          "orderNumber" : {
            "type" : "string"
          },
          "orderStatus" : {
            "type" : "string"
          },
          "orderCreateTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "refundableTickets" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RefundableTicket"
            }
          },
          "nonRefundableTickets" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/NonRefundableTicket"
            }
          },
          "totalTicketPrice" : {
            "type" : "number"
          },
          "totalRefundAmount" : {
            "type" : "number"
          },
          "totalServiceFee" : {
            "type" : "number"
          },
          "totalActualRefund" : {
            "type" : "number"
          },
          "refundRate" : {
            "type" : "string"
          },
          "applicableRule" : {
            "$ref" : "#/components/schemas/RefundRuleInfo"
          },
          "latestRefundTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "remainingSeconds" : {
            "type" : "integer",
            "format" : "int64"
          },
          "restrictions" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "suggestions" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "refundableCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "partialRefund" : {
            "type" : "boolean"
          },
          "totalFeeDisplay" : {
            "type" : "string"
          },
          "totalTicketCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalRefundDisplay" : {
            "type" : "string"
          },
          "nonRefundableCount" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "RefundRuleInfo" : {
        "type" : "object",
        "properties" : {
          "ruleId" : {
            "type" : "string"
          },
          "ruleName" : {
            "type" : "string"
          },
          "ruleDescription" : {
            "type" : "string"
          },
          "hoursBeforeDeparture" : {
            "type" : "integer",
            "format" : "int32"
          },
          "timeRangeDescription" : {
            "type" : "string"
          },
          "refundRate" : {
            "type" : "number"
          },
          "refundRateDisplay" : {
            "type" : "string"
          },
          "serviceFeeRate" : {
            "type" : "number"
          },
          "minServiceFee" : {
            "type" : "number"
          },
          "maxServiceFee" : {
            "type" : "number"
          },
          "fixedServiceFee" : {
            "type" : "number"
          },
          "applicableTrainTypes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "applicableSeatTypes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "applicableToStudent" : {
            "type" : "boolean"
          },
          "applicableToGroup" : {
            "type" : "boolean"
          },
          "allowPartialRefund" : {
            "type" : "boolean"
          },
          "allowOnlineRefund" : {
            "type" : "boolean"
          },
          "requireVerification" : {
            "type" : "boolean"
          },
          "effectiveDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expireDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "applicableCondition" : {
            "type" : "string"
          }
        }
      },
      "RefundableTicket" : {
        "type" : "object",
        "properties" : {
          "ticketNumber" : {
            "type" : "string"
          },
          "passengerName" : {
            "type" : "string"
          },
          "trainNumber" : {
            "type" : "string"
          },
          "departureTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "seatInfo" : {
            "type" : "string"
          },
          "ticketPrice" : {
            "type" : "number"
          },
          "refundAmount" : {
            "type" : "number"
          },
          "serviceFee" : {
            "type" : "number"
          },
          "actualRefund" : {
            "type" : "number"
          },
          "refundRate" : {
            "type" : "string"
          },
          "refundRule" : {
            "type" : "string"
          },
          "displayInfo" : {
            "type" : "string"
          }
        }
      },
      "ResultRefundCheckResult" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/RefundCheckResult"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "ResultListOrderInfoDTO" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OrderInfoDTO"
            }
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
    }
  }
}